scrapers/news/requirements.txt
fastapi
uvicorn[standard]
httpx
feedparser
groq
pydantic>=2
python-dotenv